# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_jamesneb_playback-backend_internal_handlers_dto.LogEntry:
    properties:
      attributes:
        additionalProperties: true
        type: object
      level:
        example: INFO
        type: string
      message:
        example: Order creation started
        type: string
      service:
        example: order-service
        type: string
      span_id:
        example: 789xyz
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
      trace_id:
        example: abc123def456
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_dto.LogsQueryResponse:
    properties:
      level:
        example: INFO
        type: string
      logs:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.LogEntry'
        type: array
      query:
        example: order failed
        type: string
      service:
        example: order-service
        type: string
      time_range:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.TimeRange'
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_dto.LogsResponse:
    properties:
      received:
        example: 10
        type: integer
      status:
        example: accepted
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_dto.TimeRange:
    properties:
      from:
        example: "2023-01-01T00:00:00Z"
        type: string
      to:
        example: "2023-01-01T23:59:59Z"
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute:
    properties:
      key:
        type: string
      value:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.AttributeValue'
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.AttributeValue:
    properties:
      boolValue:
        type: boolean
      doubleValue:
        type: number
      intValue:
        type: integer
      stringValue:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecord:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute'
        type: array
      body:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecordBody'
      droppedAttributesCount:
        type: integer
      flags:
        type: integer
      observedTimeUnixNano:
        type: integer
      severityNumber:
        type: integer
      severityText:
        type: string
      spanId:
        type: string
      timeUnixNano:
        type: integer
      traceId:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecordBody:
    properties:
      stringValue:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.LogsRequest:
    properties:
      resourceLogs:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.ResourceLog'
        type: array
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.Resource:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute'
        type: array
      droppedAttributesCount:
        type: integer
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.ResourceLog:
    properties:
      resource:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Resource'
      schemaUrl:
        type: string
      scopeLogs:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.ScopeLog'
        type: array
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.Scope:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Attribute'
        type: array
      droppedAttributesCount:
        type: integer
      name:
        type: string
      version:
        type: string
    type: object
  github_com_jamesneb_playback-backend_internal_handlers_schema.ScopeLog:
    properties:
      logRecords:
        items:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.LogRecord'
        type: array
      scope:
        $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.Scope'
    type: object
...
  internal_handlers.DownloadRequest:
    properties:
      key:
        type: string
    required:
    - key
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      message:
        example: Field validation failed
        type: string
    type: object
...
  internal_handlers.ListReplaysResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/internal_handlers.ReplayFile'
        type: array
      isTruncated:
        type: boolean
      maxKeys:
        type: integer
      nextToken:
        type: string
      totalCount:
        type: integer
    type: object
...
  internal_handlers.ReplayFile:
    properties:
      jobId:
        type: string
      key:
        type: string
      lastModified:
        type: string
      size:
        type: integer
    type: object
...
host: '{{.Host}}'
info:
  contact: {}
  description: Distributed systems event replay backend
  title: Playback Backend API
  version: '{{.Version}}'
paths:
  /api/v1/logs:
    get:
      description: Get log data for analysis
      parameters:
      - description: Service name
        in: query
        name: service
        type: string
      - description: Log level
        in: query
        name: level
        type: string
      - description: Start time (RFC3339)
        in: query
        name: from
        type: string
      - description: End time (RFC3339)
        in: query
        name: to
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.LogsQueryResponse'
      summary: Get logs
      tags:
      - logs
    post:
      consumes:
      - application/json
      description: Receive log data from OpenTelemetry
      parameters:
      - description: Log data
        in: body
        name: logs
        required: true
        schema:
          $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_schema.LogsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jamesneb_playback-backend_internal_handlers_dto.LogsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Receive logs
      tags:
      - logs
...
  /replays/download:
    post:
      consumes:
      - application/json
      description: Stream a specific Arrow IPC replay file directly from S3
      parameters:
      - description: Download request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.DownloadRequest'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Download replay file
      tags:
      - replays
  /replays/list:
    get:
      description: List available Arrow IPC replay files from S3 with pagination
      parameters:
      - description: 'Maximum number of files to return (default: 100, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: Continuation token for pagination
        in: query
        name: token
        type: string
      - description: Prefix to filter files
        in: query
        name: prefix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ListReplaysResponse'
      summary: List replay files
      tags:
      - replays
swagger: "2.0"